workers/semantics/interface-objects/003.html

<!--
onconnect = function(e) {
  var expected = [
    "ApplicationCache",
    "WorkerGlobalScope",
    "EventTarget",
    "ErrorEvent",
    "Event",
    "Worker",
    "DOMException",
    "SharedWorker",
    "MessagePort",
    "MessageEvent",
    "WorkerNavigator",
    "MessageChannel",
    "WorkerLocation",
    "Database",
    "ImageData",
    "File",
    "Blob",
    "FileList",
    "XMLHttpRequest",
    "ProgressEvent",
    "FormData",
    "ArrayBuffer",
    "Int8Array",
    "Uint8Array",
    "Uint8ClampedArray",
    "Int16Array",
    "Uint16Array",
    "Int32Array",
    "Uint32Array",
    "Float32Array",
    "Float64Array",
    "DataView",
    "CanvasProxy",
    "ImageBitmap",
    "CanvasRenderingContext2d",
    "DrawingStyle",
    "CanvasGradient",
    "CanvasPattern",
    "Path",
    "TextMetrics",
  ];
  var result = [];
  for (var i = 0; i < expected.length; ++i) {
    result.push([expected[i], expected[i] in self]);
  }
  e.ports[0].postMessage(result);
}
/*
-->
<!doctype html>
<title>available interface objects in shared worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
setup(function() {
  window.worker = new SharedWorker('#');
  worker.port.onmessage = function(e) {
    var result = e.data;
    for (var i = 0; i < result.length; ++i) {
      test(function() {
        assert_true(result[i][1]);
      }, "The " + result[i][0] + " interface object should be exposed");
    }
    done();
  }
}, {explicit_done: true});
</script>
<!--
*/
//-->
